renamed some folders for better comprehension
