+++ src/Router.tsx
@@ -1,6 +1,6 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Home from "./Pages/Home/Home";
-import Layout from "./Components/Layout/Layout";
+import Layout from "./CustomComponents/Layout/Layout";
  
 
 export default function Router() {
